extract wallet components from route
