+++ extension/src/panel/wallet/ConnectWallet.tsx
