+++ extension/src/panel/wallet/Connected.tsx
