+++ extension/src/panel/wallet/Section.tsx
