+++ extension/src/panel/wallet/SwitchChain.tsx
