+++ extension/src/panel/wallet/WalletDisconnected.tsx
