+++ extension/src/panel/wallet/index.ts
