+++ extension/src/panel/wallet/injectedWallet/InjectedWallet.tsx
