+++ extension/src/panel/wallet/injectedWallet/InjectedWalletConnect.tsx
