+++ extension/src/panel/wallet/injectedWallet/index.ts
