+++ extension/src/panel/wallet/providerLogo/ProviderLogo.tsx
