+++ extension/src/panel/wallet/providerLogo/index.ts
