+++ extension/src/panel/wallet/providerLogo/wallet-connect-logo.png
