+++ extension/src/panel/wallet/walletConnect/WalletConnect.tsx
