+++ extension/src/panel/wallet/walletConnect/WalletConnectConnect.tsx
