+++ extension/src/panel/wallet/walletConnect/index.ts
