+++ extension/tsconfig.json
@@ -27,6 +27,7 @@
       "@/state": ["./src/panel/state/index.tsx"],
       "@/providers": ["./src/panel/providers/index.ts"],
       "@/providers-ui": ["./src/panel/providers-ui/index.ts"],
+      "@/wallet": ["./src/panel/wallet/index.ts"],
       "@/inject-bridge": ["./src/inject/bridge/index.ts"],
       "@/e2e-utils": ["./e2e/utils/index.ts"],
       "@/test-utils": ["./test-utils/index.ts"],
